Work around empty go.o file.
On Windows the go.o file may be empty when the 'ar' command is run to
generate a C-style library archive. This fix replaces all Sprintf
path generation with calls to the canonical path.Join in lib.go, and
also adds an os.Stat(). This appears to force Windows to flush the
contents of go.o in a way that makes them visible to the 'ar' process
which runs directly after.

It is admittedly a hack, but an effective one. Perhaps a deeper
investigation into the matter will reveal something useful.